+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 ### Changes
+* Apples now only drop from leaves in Spring
 * Foxhounds now have a really high life span
 * Lowered the amount of phantom membrane required to make and repair a hang glider
 * Increased low tier materials durability